+++ src/Import/MsfsPln.ts
@@ -16,6 +16,11 @@ export class MsfsPln extends GarminFpl {
   read(configFileContent: string): void {
     const waypointTableXml = this.getXmlNode(configFileContent, "FlightPlan.FlightPlan");
 
+    const versionId = Number(this.getXmlNode(waypointTableXml, "AppVersionMajor"));
+    if (versionId <= 0 || versionId > 12) {
+      throw Error("Unknown flight plan version ID");
+    }
+
     this.cruisingAlt = Number(this.getXmlNode(waypointTableXml, "CruisingAlt"));
     const waypointsXml = this.getXmlNodes(waypointTableXml, "ATCWaypoint");
     this.waypoints = waypointsXml.map((xml, index): GaminFplWaypoint => {